Separation of functionality
- fix counter bugs
- add key to env
