+++ .env
@@ -5,5 +5,6 @@ MONITOR_BACKEND_PATH_SOCKETIO=/hkube/monitor-server/socket.io
 MONITOR_BACKEND_HOST=dev1.hkube.io
 MONITOR_BACKEND_PORT=80
 MONITOR_BACKEND_PATH=/hkube/monitor-server 
+REACT_APP_BACKEND_HOST=dev1.hkube.io
