+++ src/graphql/config.js
@@ -3,7 +3,7 @@ import { createHttpLink, ApolloClient } from '@apollo/client';
 import cache from 'cache';
 
 const httpLink = createHttpLink({
-  uri: 'https://dev1.hkube.io/hkube/api-server/graphql',
+  uri: `https://${process.env.REACT_APP_BACKEND_HOST}/hkube/api-server/graphql`,
 });
 
 export const apolloClient = new ApolloClient({
